+++ views/config/index.php
@@ -104,6 +104,7 @@ $module = Yii::$app->getModule('clean-theme');
             <?= $form->field($model, 'fontFamily')->textInput() ?>
             <?= $form->field($model, 'fontSize')->textInput(['type' => 'number', 'step' => 1, 'min' => 6]) ?>
             <?= $form->field($model, 'fontWeight')->textInput() ?>
+            <?= $form->field($model, 'fontBoldWeight')->textInput() ?>
             <?= $form->endCollapsibleFields() ?>
 
             <?= $form->beginCollapsibleFields(Yii::t('CleanThemeModule.config', 'Heading fonts')) ?>