+++ docs/CHANGELOG.md
@@ -11,6 +11,8 @@ Unreleased (wait for DeviceDetectorHelper::getBodyClasses() to be available)
 - Enh: Topbar brand text not centered vertically
 - Enh: When hovering a User or Space, add a black translucent overlay to make the white text visible if the image is light
 - Enh: Load all Google Font weights
+- Enh: Add configuration for Bold font (to allow changing the weight)
+- Fix: Safe area padding on mobile app
 
 2.2.0 (November 26, 2024)
 --------------------